// src/lib/form-engine/types.ts
export type FormTemplateWithSections = FormTemplate & {
  sections: FormSectionWithQuestions[];
};

export type FormSectionWithQuestions = FormSection & {
  questions: FormQuestionWithDetails[];
};
// ...
export type FormQuestionWithDetails = Omit<FormQuestion, 'repeatableConfig'> & {
  repeatableConfig: Prisma.JsonValue | null;
  options: QuestionOption[];
  scoringConfig: ScoringConfig | null;
  dictionary: QuestionDictionary | null;
};